refactor: move the resume components to a separate folder.
